Before CSS commit
Every intended feature works as intended . This will be the final stage for the technical side of the web server.
Next commit will only include the CSS files to make everything look pretty.

On REST api there are 3 folders for People, Grade, Subject which contain the handling logic of the http requests.

In React project there are 4 pages Students, Teachers, Grades, Subjects with 4 corresponding folders that will be requesting data manipulation for each of the pages.